Extracted section and convert items stuff into new Section class
